Allow empty bytecodes in Contract.from_abi
Some chains have decided to add ABI-driven precompiled contracts,
similar to the current cryptographic precompiles available in Ethereum.
These precompiles are handled directly by the file `contracts.go` and
as such do not have an available bytecode. However, it is possible to
interact with them if the function selectors are available.

Example chains include (1) abeychain, (2) Moonbeam, both of which use a
staking precompile with no bytecode and just an ABI. See also
harmony-one/harmony#3906